feat(website): handle styles not appearing in root component
This workaround
for this bug https://github.com/remix-run/remix/issues/1136.
