+++ apps/website/app/lib/routes.d.ts
@@ -1,28 +1,28 @@
 declare module "routes-gen" {
   export type RouteParams = {
+    "/": Record<string, never>;
     "/questions/create-or-update": Record<string, never>;
     "/information/toolchain": Record<string, never>;
     "/questions/solve/:slug": { "slug": string };
     "/playground": Record<string, never>;
     "/classes/create": Record<string, never>;
     "/questions/list": Record<string, never>;
-    "/classes/:id": { "id": string };
-    "/": Record<string, never>;
-    "/auth": Record<string, never>;
     "/auth/logout": Record<string, never>;
+    "/classes/:id": { "id": string };
+    "/auth": Record<string, never>;
   };
 
   export function route<
     T extends
+      | ["/"]
       | ["/questions/create-or-update"]
       | ["/information/toolchain"]
       | ["/questions/solve/:slug", RouteParams["/questions/solve/:slug"]]
       | ["/playground"]
       | ["/classes/create"]
       | ["/questions/list"]
-      | ["/classes/:id", RouteParams["/classes/:id"]]
-      | ["/"]
-      | ["/auth"]
       | ["/auth/logout"]
+      | ["/classes/:id", RouteParams["/classes/:id"]]
+      | ["/auth"]
   >(...args: T): typeof args[0];
 }
