+++ apps/website/app/routes/__boundary/index.tsx
@@ -2,7 +2,6 @@ import { json } from '@remix-run/node';
 import { Link, useLoaderData } from '@remix-run/react';
 import { route } from 'routes-gen';
 import { requireValidJwt } from '~/lib/services/auth.server';
-
 import type { LoaderArgs } from '@remix-run/node';
 
 export const loader = async ({ request }: LoaderArgs) => {
