--- apps/website/app/routes/auth.tsx
@@ -1,5 +1,0 @@
-import { Outlet } from '@remix-run/react';
-
-export default () => {
-  return <Outlet />;
-};
